requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.5
openai==0.28.1
requests==2.28.2
python-dotenv==1.0.0
pytest==8.3.5
pydantic-settings==2.0.3
pandas==2.1.1
numpy==1.26.1
matplotlib==3.8.0
seaborn==0.13.0
tqdm==4.67.0
scikit-learn==1.3.2